project/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
Django==5.2.3
django-allauth==65.9.0
django-crispy-forms==2.4
django-environ==0.12.0
sqlparse==0.5.3
psycopg2-binary==2.9.10
uWSGI
